feat(painter): name Painterly's settings for what they do
The stroke sizes were fractions of the image on a 0-1 slider, labelled
"Largest Brush" beside the Brush menu, with the useful range in the
first quarter of the track. They are now Largest and Smallest Stroke,
percentages of the image with soft ranges around the defaults. Steps
become Passes and the opacities First and Last Pass Opacity. Coverage
and Edge Threshold are percentages, and Edge Threshold moves to a new
Placement group with Coverage, since it decides where strokes land.
Coverage now runs from 1 % to 200 %.

Smoothing was the one setting left in texels of the built image, so
raising the resolution repainted the picture: on a 4K test picture the
stroke directions at 4096 were a median 25.6 degrees from those at
2048. It is now in texels of a 2048 image and scales with the
resolution, which brings that to 2.8 degrees.

Random Rotation and Rotation Range merge into one angle, since the
settings panel cannot hide a range that does nothing while its switch
is off. An angle of 0 paints exactly what the switch off did.

plan.Settings keeps v2's quantities and Settings.of converts, rounding
each percentage to a single-precision fraction as v2's properties were,
so the stamp counts stay exact. At 2048 the default settings, and a set
of changed ones, paint the same pixels as before. Because the
fingerprint records the blur in texels, a new resolution changes the
settings part too, so the out-of-date reason checks the resolution
first.

+++ filters/painter/build.py
@@ -273,7 +273,7 @@ def _run(pool, spec: FilterSpec, texture, params: dict | None = None, *, keep: b
     return result
 
 
-def _blurred(pool, texture, sigma: int, *, keep: bool, progress: tuple[str, float]):
+def _blurred(pool, texture, sigma: float, *, keep: bool, progress: tuple[str, float]):
     """*texture* through v2's gaussian: one pass per axis, cut off at two sigma.
 
     A generator yielding *progress* between the two passes, each of
